refactor(packaging): change file names and structure
