+++ src/components/TrafficLight/TrafficLight.css
